feat(query): Increase phrase slop to 5
Pelias currently indexes all addresses with the format '[housenumber]
[streetname]'. This is the predominate style in the USA, but many
countries, especially in Europe use '[streetname] [housenumber]'.

Either style can be matched, as long as the total number of tokens in
the housenumber and streetname are less than the 'slop' value for our
Elasticsearch `match_phrase` queries.

The current value of 4 supports _most_ cases, but there are a good
number of street names with 4 tokens, requiring a slop size of 5.

With this change we should handle an even higher percentage of streets,
hopefully with minimal downsides.